2.6.3
- Updated `ScalarMult` method in `X25519.cs` to use `System.Int16` for loop variable.
- Changed `Localizer` and `MultiLocalizer` classes to allow inheritance and made several methods virtual.
- Added comprehensive test classes for `ChaCha20Poly1305`, `Ed25519`, `Srp6`, and `X25519` algorithms.
- Created hashing tests for SHA1, SHA224, and SHA256.
- Implemented tests for symmetric encryption algorithms: Arc4, Blowfish, ChaCha20, Salsa20, and Speck.
- Added `LiteSerializer` tests for arrays, objects, primitives, and strings.
- Updated project files to include necessary references and settings.
- Modified solution file to include new test projects.

+++ src/Nalix.Cryptography/Asymmetric/X25519.cs
@@ -71,7 +71,7 @@ public static unsafe class X25519
         z3[0] = 1;
 
         System.UInt32 swap = 0;
-        for (System.Byte pos = 254; pos >= 0; --pos)
+        for (System.Int16 pos = 254; pos >= 0; --pos)
         {
             System.UInt32 b = (System.UInt32)(scalar[pos / 8] >> (pos & 7)) & 1;
             swap ^= b;
